ml/requirements.txt
scikit-learn>=1.5.0
pandas>=2.2.0
numpy>=1.26.0
joblib>=1.3.2
psycopg2-binary>=2.9.9
matplotlib>=3.8.0
flask>=3.0.0
flask-cors>=4.0.0
prophet>=1.1.5
python-dotenv>=1.0.0
scipy>=1.11.0
seaborn>=0.13.0